requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23

# ML & Data Processing
torch==2.1.1
transformers==4.35.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Image & Video Processing
opencv-python==4.8.1.78
Pillow==10.1.0
deepface==0.0.79
ffmpeg-python==0.2.0

# Web Scraping & API
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# Database
aiosqlite==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0 